backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "VerifID"
requires-python = ">=3.14"

dependencies = [
    "fastapi[standard]==0.135.2",
    "pydantic[email]",
    "pydantic-settings",
    "sqlmodel==0.0.37",
    "sqlalchemy[asyncio]",
    "alembic",
    "python-jose[cryptography]",
    "python-multipart",
    "msal>=1.35.1",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pre-commit",
]

[tool.setuptools]
packages = ["backend"]

# ── Ruff (linter PEP8) ──────────────────────────────────────────────────────
[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

# ── Ruff — format ────────────────────────────────────────────────────────────
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

# ── Mypy (type checking) ─────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.14"
strict = true
ignore_missing_imports = true
